Task2

+++ app/src/main/java/francisco/simon/myfinance/ui/features/account/mapper/Mapper.kt
@@ -7,7 +7,7 @@ fun Account.toAccountUI(): AccountUI {
     return AccountUI(
         id = id,
         name = name,
-        balance = balance,
+        balance = balance.toBigDecimal(),
         currency = currency
     )
 }
